updated for Viewing Transaction

+++ W4D5_Project/KanenOnlineShopping/project/KanenOnlineShopping/src/com/oocl/mnlbc/group3/model/OrderBean.java
@@ -16,14 +16,14 @@ public class OrderBean implements Serializable {
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
-	private int orderId;
+	private long orderId;
 	private int userId;
 	private String orderDate;
 	private double totalCost;
 	private String orderStatus;
 	private List<ItemsBean> items;
 
-	public OrderBean(int orderId, int userId, String orderDate, double totalCost, String orderStatus,
+	public OrderBean(long orderId, int userId, String orderDate, double totalCost, String orderStatus,
 			List<ItemsBean> items) {
 		super();
 		this.orderId = orderId;
@@ -37,7 +37,7 @@ public class OrderBean implements Serializable {
 	/**
 	 * @return the orderId
 	 */
-	public int getOrderId() {
+	public long getOrderId() {
 		return orderId;
 	}
